Change unpack_trees' 'reset' flag into an enum
Traditionally, unpack_trees_options->reset was used to signal that it
was okay to delete any untracked files in the way.  This was used by
`git read-tree --reset`, but then started appearing in other places as
well.  However, many of the other uses should not be deleting untracked
files in the way.  Change this value to an enum so that a value of 1
(i.e. "true") can be split into two:
   UNPACK_RESET_PROTECT_UNTRACKED,
   UNPACK_RESET_OVERWRITE_UNTRACKED
In order to catch accidental misuses, define with the enum a special
value of
   UNPACK_RESET_INVALID = 1
which will trigger a BUG().

Modify existing callers so that
   read-tree --reset
   reset --hard
   checkout --force
continue using the UNPACK_RESET_OVERWRITE_UNTRACKED logic, while other
callers, including
   am
   checkout without --force
   stash  (though currently dead code; reset always had a value of 0)
   numerous callers from rebase/sequencer to reset_head()
will use the new UNPACK_RESET_PROTECT_UNTRACKED value.

In order to protect untracked files but still allow deleting of ignored
files, we also have to setup unpack_trees_opt.dir.  It may make sense to
set up unpack_trees_opt.dir in more cases, but here I tried to only do
so in cases where we switched from deleting all untracked files to
avoiding doing so (i.e. where we now use
UNPACK_RESET_PROTECT_UNTRACKED).

Also, note that 'git checkout <pathspec>' currently also allows
overwriting untracked files.  That case should also be fixed, but it
does not use unpack_trees() and thus is outside the scope of the current
changes.

Signed-off-by: Elijah Newren <newren@gmail.com>