diff: --quote-path-with-sp
Long time ago, we had a discussion with GNU patch/diff maintainer
and agreed that pathnames with certain "difficult" bytes needs to be
quoted to ensure the resulting patch is machine parseable in an
unambiguous way [*1*].  Recently, we saw a report that found that
GNU patch is unhappy with our diff output for a path with SP in it
[*2*].

Teach "git diff" and friends the "--quote-path-with-sp" option, that
encloses a pathname with SP in it inside a pair of double-quotes,
even though there is otherwise no byte in the pathname that need to
be encoded in the octal.

As an earlier parts of t/t3902 (outside the patch context) shows,
output from "ls-files", "ls-tree", and "diff --name-only" all follow
the same rule to decide paths with what bytes in them need quoting
and how they are quoted.

This experimental option deliberately refrains from touching these
output and affects ONLY the paths that appear in the patch header,
i.e. "diff --git", "--- a/path" and "+++ b/path" lines, that GNU
patch may care.  This is to minimize potential damage this change
may cause to tools and scripts the users have been relying on.

 *1* https://lore.kernel.org/git/87ek6s0w34.fsf@penguin.cs.ucla.edu/
 *2* https://lore.kernel.org/git/YR9Iaj%2FFqAyCMade@tilde.club/

Signed-off-by: Junio C Hamano <gitster@pobox.com>