sparse-checkout: clear patterns when switching modes
Previously, when a user runs 'git sparse-checkout init --cone', the
existing patterns would remain, even if the patterns were not in cone
mode. This causes confusion as to how the patterns should work when
later 'git sparse-checkout add' commands append to the pattern list.

In fact, the way these patterns were modified was not even a strict
appending of patterns, but mutating and reordering patterns because of
how the paths were interpreted and rewritten.

As a first step, we shall start overwriting the pattern set completely
when switching to cone mode, unless the existing patterns already match
cone mode patterns. The 'use_cone_patterns' member is set to false if
the patterns fail to parse using cone mode restrictions.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>