amend error message violations
Issue no #635 in the gitgitgadget/git repo has a brief description
about the violation of error message style convention. These
violations are - 1) Do not end error messages with a full stop. 2) Do
not capitalize ("unable to open %s", not "Unable to open %s") 3) Say
what the error is first ("cannot open %s", not "%s: cannot open")

This commit is the first commit to amend those error strings. It
covers files in the root directory that start with letter 'a'. This is
a small patch to start as suggested. There will be more patches to
cover all files and directories.

One thing to note that I didn't change any .po file as there is a
doubt whether those are good to change or not.

Signed-off-by: Abhradeep Chakraborty <chakrabortyabhradeep79@gmail.com>